backend/requirements.txt
twelvelabs
weaviate-client
flask
flask-cors
boto3
python-dotenv
requests
gunicorn
apscheduler

